Rename jmrit.roco z21* classes to Z21*
"Initial-cap" is our convention for class names. Used NetBeans 8.0.2 to do migration.
