+++ java/src/jmri/jmrix/roco/z21/simulator/ConnectionConfig.java
@@ -1,4 +1,3 @@
-// ConnectionConfig.java
 package jmri.jmrix.roco.z21.simulator;
 
 import javax.swing.JPanel;
@@ -8,13 +7,12 @@ import javax.swing.JPanel;
  * Handle configuring an z21 layout connection via a z21Simulator
  * adapter.
  * <P>
- * This uses the {@link z21SimulatorAdapter} class to do the actual connection.
+ * This uses the {@link Z21SimulatorAdapter} class to do the actual connection.
  *
  * @author Bob Jacobsen Copyright (C) 2001, 2003
  * @author Paul Bender Copyright (C) 2009
- * @version	$Revision$
  *
- * @see z21SimulatorAdapter
+ * @see Z21SimulatorAdapter
  */
 public class ConnectionConfig extends jmri.jmrix.roco.z21.ConnectionConfig {
 
@@ -41,7 +39,7 @@ public class ConnectionConfig extends jmri.jmrix.roco.z21.ConnectionConfig {
     @Override
     protected void setInstance() {
         if (adapter == null) {
-            adapter = new z21SimulatorAdapter();
+            adapter = new Z21SimulatorAdapter();
         }
     }
