+++ java/src/jmri/jmrix/roco/z21/simulator/configurexml/ConnectionConfigXml.java
@@ -1,21 +1,20 @@
 package jmri.jmrix.roco.z21.simulator.configurexml;
 
 import jmri.jmrix.roco.z21.simulator.ConnectionConfig;
-import jmri.jmrix.roco.z21.simulator.z21SimulatorAdapter;
+import jmri.jmrix.roco.z21.simulator.Z21SimulatorAdapter;
 
 /**
- * Handle XML persistance of layout connections by persistening the
- * z21SimulatorAdapter (and connections). Note this is named as the XML version
- * of a ConnectionConfig object, but it's actually persisting the
- * z21SimulatorAdapter.
- * <P>
+ * Handle XML persistence of layout connections by persisting the
+ Z21SimulatorAdapter (and connections). Note this is named as the XML version
+ of a ConnectionConfig object, but it's actually persisting the
+ Z21SimulatorAdapter.
+ <P>
  * This class is invoked from jmrix.JmrixConfigPaneXml on write, as that class
  * is the one actually registered. Reads are brought here directly via the class
  * attribute in the XML.
  *
  * @author Bob Jacobsen Copyright: Copyright (c) 2003
  * @author Paul Bender Copyright: Copyright (c) 2009
- * @version $Revision$
  */
 public class ConnectionConfigXml extends jmri.jmrix.roco.z21.configurexml.ConnectionConfigXml {
 
@@ -26,7 +25,7 @@ public class ConnectionConfigXml extends jmri.jmrix.roco.z21.configurexml.Connec
     @Override
     protected void getInstance() {
         if (adapter == null) {
-            adapter = new z21SimulatorAdapter();
+            adapter = new Z21SimulatorAdapter();
         }
     }
