+++ java/src/jmri/jmrix/roco/z21/swing/packetgen/PacketGenAction.java
@@ -1,11 +1,3 @@
-/**
- * PacketGenAction.java
- *
- * Description:	Swing action to create and register a z21 PacketGenFrame object
- *
- * @author	Bob Jacobsen Copyright (C) 2001, 2002
- * @version	$Revision$
- */
 package jmri.jmrix.roco.z21.swing.packetgen;
 
 import java.awt.event.ActionEvent;
@@ -13,20 +5,21 @@ import javax.swing.AbstractAction;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+/**
+ * Swing action to create and register a z21 PacketGenFrame object
+ *
+ * @author	Bob Jacobsen Copyright (C) 2001, 2002
+ */
 public class PacketGenAction extends AbstractAction {
 
-    /**
-     *
-     */
-    private static final long serialVersionUID = 2121741838451012843L;
-    jmri.jmrix.roco.z21.z21SystemConnectionMemo _memo = null;
+    jmri.jmrix.roco.z21.Z21SystemConnectionMemo _memo = null;
 
-    public PacketGenAction(String s, jmri.jmrix.roco.z21.z21SystemConnectionMemo memo) {
+    public PacketGenAction(String s, jmri.jmrix.roco.z21.Z21SystemConnectionMemo memo) {
         super(s);
         _memo = memo;
     }
 
-    public PacketGenAction(jmri.jmrix.roco.z21.z21SystemConnectionMemo memo) {
+    public PacketGenAction(jmri.jmrix.roco.z21.Z21SystemConnectionMemo memo) {
         this("Send Z21 Message", memo);
     }
 
@@ -35,7 +28,7 @@ public class PacketGenAction extends AbstractAction {
         // If there is no system memo given, assume the system memo
         // is the first one in the instance list.
         _memo = jmri.InstanceManager.
-                getList(jmri.jmrix.roco.z21.z21SystemConnectionMemo.class).get(0);
+                getList(jmri.jmrix.roco.z21.Z21SystemConnectionMemo.class).get(0);
     }
 
     public PacketGenAction() {
@@ -57,6 +50,3 @@ public class PacketGenAction extends AbstractAction {
     }
     private final static Logger log = LoggerFactory.getLogger(PacketGenAction.class.getName());
 }
-
-
-/* @(#)LocoGenAction.java */
