+++ java/src/jmri/jmrix/roco/z21/swing/packetgen/PacketGenFrame.java
@@ -1,14 +1,12 @@
-// PacketGenFrame.java
 package jmri.jmrix.roco.z21.swing.packetgen;
 
-import jmri.jmrix.roco.z21.z21Message;
-import jmri.jmrix.roco.z21.z21TrafficController;
+import jmri.jmrix.roco.z21.Z21Message;
+import jmri.jmrix.roco.z21.Z21TrafficController;
 
 /**
  * Frame for user input of XpressNet messages
  *
  * @author	Bob Jacobsen Copyright (C) 2001,2002
- * @version	$Revision$
  */
 public class PacketGenFrame extends jmri.jmrix.swing.AbstractPacketGenFrame {
 
@@ -32,19 +30,19 @@ public class PacketGenFrame extends jmri.jmrix.swing.AbstractPacketGenFrame {
         tc.sendz21Message(createPacket(packetTextField.getSelectedItem().toString()), null);
     }
 
-    z21Message createPacket(String s) {
+    Z21Message createPacket(String s) {
         if (s.equals("")) {
             return null; // message cannot be empty
         }
-        z21Message m = new z21Message(s);
+        Z21Message m = new Z21Message(s);
         return m;
     }
 
     // connect to the TrafficController
-    public void connect(z21TrafficController t) {
+    public void connect(Z21TrafficController t) {
         tc = t;
     }
 
     // private data
-    private z21TrafficController tc = null;
+    private Z21TrafficController tc = null;
 }
