+++ java/test/jmri/jmrix/roco/PackageTest.java
@@ -1,4 +1,3 @@
-// PackageTest.java
 package jmri.jmrix.roco;
 
 import junit.framework.Test;
@@ -9,7 +8,6 @@ import junit.framework.TestSuite;
  * Tests for the jmri.jmrix.roco package
  *
  * @author	Bob Jacobsen
- * @version	$Revision$
  */
 public class PackageTest extends TestCase {
 
@@ -29,7 +27,7 @@ public class PackageTest extends TestCase {
     public static Test suite() {
         TestSuite suite = new TestSuite("jmri.jmrix.roco.RocoTest");  // no tests in this class itself
         suite.addTest(new TestSuite(RocoConnectionTypeListTest.class));
-        suite.addTest(jmri.jmrix.roco.z21.z21Test.suite());
+        suite.addTest(jmri.jmrix.roco.z21.Z21Test.suite());
 
         return suite;
     }
